pyproject.toml
[project]
name = "nonebot-plugin-paminet-nodirtymsg"
version = "0.2.0"
description = "基于Trie树的高效群聊违禁词过滤插件,自动撤回群成员的敏感发言.需要管理员权限."
authors = [
    { name = "Mipa", email = "shcdwrzh@outlook.com" }
]
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
keywords = ["nonebot", "chat", "moderation", "filter"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Communications :: Chat",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "nonebot2>=2.0.0",
    "nonebot-adapter-onebot>=2.0.0",
    "pydantic>=1.10.0",
    "aiofiles>=23.1.0",
]

[project.urls]
Homepage = "https://github.com/PamiNET-Corp/nonebot-plugin-paminet-nodirtymsg"
Repository = "https://github.com/PamiNET-Corp/nonebot-plugin-paminet-nodirtymsg"
Documentation = "https://github.com/PamiNET-Corp/nonebot-plugin-paminet-nodirtymsg#readme"
Issues = "https://github.com/PamiNET-Corp/nonebot-plugin-paminet-nodirtymsg/issues"

[project.entry-points."nonebot.plugins"]
paminet_nodirtymsg = "nonebot_plugin_paminet_nodirtymsg"

[tool.setuptools]
packages = ["nonebot_plugin_paminet_nodirtymsg"]
include-package-data = true

[tool.setuptools.package-data]
nonebot_plugin_paminet_nodirtymsg = ["data/*.json"]